Created teh PopularItemsBehaviour script
This scripp will control the items that are displayed, Right now it is using hard code test data. But it will later use a query to the database, Also changed the Item class to ahave IComparable, This is for the Sort method. Which will make the list sort by the numberOfClicks this will make it easier the place the items into the Popular Items Panel